+++ fbw-a380x/src/systems/instruments/src/MFD/FMC/FmcAircraftInterface.ts
@@ -323,6 +323,7 @@ export class FmcAircraftInterface {
 
     const taxiFuel = this.fmgc.data.taxiFuel.get() ?? 0;
     const tow = (this.fmc.fmgc.getGrossWeightKg() ?? maxGw) - (this.fmgc.isAnEngineOn() ? 0 : taxiFuel);
+    const flapConf = this.fmgc.data.takeoffFlapsSetting.get();
 
     return (
       (this.flightPlanService.active.performanceData.v1 ?? Infinity) < Math.trunc(A380SpeedsUtils.getVmcg(zp)) ||
@@ -330,7 +331,7 @@ export class FmcAircraftInterface {
       (this.flightPlanService.active.performanceData.v2 ?? Infinity) < Math.trunc(1.1 * A380SpeedsUtils.getVmca(zp)) ||
       (Number.isFinite(tow) &&
         (this.flightPlanService.active.performanceData.v2 ?? Infinity) <
-          Math.trunc(1.13 * A380SpeedsUtils.getVs1g(tow / 1000, this.fmgc.data.takeoffFlapsSetting.get(), true)))
+          Math.trunc(1.13 * A380SpeedsUtils.getVs1g(tow, flapConf > 1 ? flapConf + 1 : flapConf, true)))
     );
   }
 
@@ -867,7 +868,9 @@ export class FmcAircraftInterface {
     if (!Vtap) {
       // Overspeed protection
       const vMax = this.speedVmax.get();
-      Vtap = Math.min(this.managedSpeedTarget ?? Vmo - 5, vMax - 5);
+      const greenDot = this.fmgc.data.greenDotSpeed.get();
+      const vMin = SimVar.GetSimVarValue('L:A32NX_FLAPS_HANDLE_INDEX', 'Number') === 0 && greenDot ? greenDot : 0;
+      Vtap = Math.min(Math.max(this.managedSpeedTarget ?? Vmo - 5, vMin), vMax - 5);
     }
     SimVar.SetSimVarValue('L:A32NX_SPEEDS_MANAGED_PFD', 'knots', vPfd);
     SimVar.SetSimVarValue('L:A32NX_SPEEDS_MANAGED_ATHR', 'knots', Vtap);
@@ -1158,7 +1161,7 @@ export class FmcAircraftInterface {
 
     // Calculate approach speeds. Independent from ADR data
     const approachSpeeds = new A380OperatingSpeeds(
-      ldgWeight / 1000,
+      ldgWeight,
       0,
       this.fmgc.data.approachFlapConfig.get(),
       FmgcFlightPhase.Approach,
@@ -1187,7 +1190,7 @@ export class FmcAircraftInterface {
         flaps = 5; // CONF 1
       }
       const speeds = new A380OperatingSpeeds(
-        grossWeight / 1000,
+        grossWeight,
         cas,
         flaps,
         this.flightPhase.get(),